refactor: syntax, imports location
